refactor: 月曜日の日付を取得する関数をutils配下に移動
