+++ src/components/newChart/LayoutIcon/LayoutIcon.module.css
@@ -9,7 +9,6 @@
 .gridOn,
 .gridOff {
   margin: 10px 5px;
-
   font-size: 3.5rem;
 }
