+++ src/components/newChart/SelectDate/SelectDate.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import { SelectDate } from "./SelectDate";
 import "@testing-library/jest-dom";
-import { GETMONDAYOFLASTWEEK, GETMONDAYOFTHISWEEK } from "@/constants/constant";
+import { GETMONDAYOFLASTWEEK, GETMONDAYOFTHISWEEK } from "@/utils/getMonday";
 
 const handleLastClick = jest.fn();
 const handleThisClick = jest.fn();
@@ -9,14 +9,20 @@ const handleThisClick = jest.fn();
 describe("LayoutIconコンポーネントの単体テスト", () => {
   test("レイアウト変更アイコンが表示される", () => {
     const options = { month: "2-digit" as const, day: "2-digit" as const };
-    const lastMondayDate = GETMONDAYOFLASTWEEK.toLocaleDateString(undefined, options);
-    const thisMondayDate = GETMONDAYOFTHISWEEK.toLocaleDateString(undefined, options);
+    const lastMondayDate = GETMONDAYOFLASTWEEK.toLocaleDateString(
+      undefined,
+      options
+    );
+    const thisMondayDate = GETMONDAYOFTHISWEEK.toLocaleDateString(
+      undefined,
+      options
+    );
     render(
       <SelectDate
         weekCheck="all"
         handleLastClick={handleLastClick}
         handleThisClick={handleThisClick}
-      />,
+      />
     );
 
     const lastMonday = screen.getByText("先週");
@@ -34,7 +40,7 @@ describe("LayoutIconコンポーネントの単体テスト", () => {
         weekCheck="all"
         handleLastClick={handleLastClick}
         handleThisClick={handleThisClick}
-      />,
+      />
     );
     const lastMonday = screen.getByText("先週");
     const thisMonday = screen.getByText("今週");
@@ -48,7 +54,7 @@ describe("LayoutIconコンポーネントの単体テスト", () => {
         weekCheck="last"
         handleLastClick={handleLastClick}
         handleThisClick={handleThisClick}
-      />,
+      />
     );
     const lastMonday = screen.getByText("先週");
     const thisMonday = screen.getByText("今週");
@@ -62,7 +68,7 @@ describe("LayoutIconコンポーネントの単体テスト", () => {
         weekCheck="this"
         handleLastClick={handleLastClick}
         handleThisClick={handleThisClick}
-      />,
+      />
     );
     const lastMonday = screen.getByText("先週");
     const thisMonday = screen.getByText("今週");
