+++ src/constants/constant.ts
@@ -18,21 +18,3 @@ export const GENRE_ARTISTS = [
   { id: 95, name: "キッズ" },
   { id: 173, name: "映画/ゲーム" },
 ];
-
-const getMondayOfThisWeek = () => {
-  const today = new Date();
-  const dayOfWeek = today.getDay();
-  const diff = dayOfWeek === 0 ? -6 : 1 - dayOfWeek;
-  const monday = new Date(today.setDate(today.getDate() + diff));
-  return monday;
-};
-export const GETMONDAYOFTHISWEEK = getMondayOfThisWeek();
-
-const getMondayOfLastWeek = () => {
-  const today = new Date();
-  const dayOfWeek = today.getDay();
-  const diff = dayOfWeek === 0 ? -6 : 1 - dayOfWeek;
-  const monday = new Date(today.setDate(today.getDate() + diff - 7));
-  return monday;
-};
-export const GETMONDAYOFLASTWEEK = getMondayOfLastWeek();
